Splitting up
Split MaterialDatabase up into MaterialDatabase and Material with separate files.
Also split up GlazeDatabase into GlazeDatabase, Glaze, and PyrometricCone with separate files.
Changed ErrorPopup to use a for loop for its row creation, so it would be more modular in future if I choose to repurpose the class.
Altered the molecule classes to use references where possible, to minimize overhead. Plan to do the same anywhere I create large databases and move them around.
