+++ GlazeBuilder/MaterialDatabase.cs
@@ -87,16 +87,4 @@ namespace GlazeBuilder
         }
         */
     }
-
-    class Material
-    {
-        public Material(JProperty material_json_property)
-        {
-            Name = material_json_property.Name;
-        }
-
-        string Name { get; set; }
-        List<CompoundMolecule> CompoundMolecules { get; set; }
-        List<SimpleMolecule> SimpleMolecules { get; set; }
-    }
 }
